+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 210:
 
 * Tidy up read implementation
 * Fix stable_async_op_base javadoc
+* Better handling of stream timeouts
 
 --------------------------------------------------------------------------------